[refactor] packaging
